PR feedback, break up unit tests
